requirements.txt
contourpy==1.3.0
cycler==0.12.1
dlib==19.24.6
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pillow_heif==0.20.0
pyparsing==3.2.0
